commit 1. fans num schedule and usr fans ,usr follow ,follow/unfollow

+++ common/src/main/java/mine/cloud/DMicro/pojo/User.java
@@ -22,5 +22,5 @@ public class User {
 
     private String usrPassword;
 
-    private Integer usrFans;
+    private Long usrFans;
 }
